setup-flask/requirements.txt
flask
python-dotenv
authlib